requirements.txt
django
gunicorn
django-heroku
python-dotenv
django-allauth
